requirements.txt
Babel==2.12.1
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
celery[redis]
redis
flask-security-too
flask_restful
flask_cors
flask_caching
bleach
mypy
flask_restful
bcrypt
pandas
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
passlib==1.7.4
PyJWT==2.6.0
pytz==2022.7.1
speaklater==1.3
SQLAlchemy==2.0.6
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
